+++ crates/burn-jit/src/fusion/on_write/mod.rs
@@ -2,7 +2,6 @@ pub(crate) mod builder;
 pub(crate) mod io;
 pub(crate) mod ir;
 pub(crate) mod kernel;
-pub(super) mod position;
+pub(crate) mod settings;
 
 pub mod trace;
-pub(crate) mod trace_builder;
