+++ crates/burn-jit/src/fusion/on_write/trace/mod.rs
@@ -1,0 +1,14 @@
+pub(crate) mod executor;
+pub(crate) mod inputs;
+pub(crate) mod outputs;
+pub(crate) mod vectorization;
+
+mod base;
+mod builder;
+mod plan;
+mod runner;
+
+pub use base::*;
+pub use builder::*;
+pub use plan::*;
+pub use runner::*;
